refactor: adiciona pasta apiRequest
